requirements.txt
crewai==0.28.8
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
streamlit==1.28.1
pydantic==2.5.0
